chore: rename plot function file
